+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2022-11-16 22:46:05
 
 - Add support for saving the image prompt and associated mask
+- Add support for greyscale image filtering - Chrome only - no safari
 
 ## 2022-11-15 21:41:50